update stop script

+++ stopapp.sh
@@ -1,4 +1,10 @@
 #!/usr/bin/env bash
 echo 'Starting Spring Boot app'
 cd '/home/ec2-user'
-kill -9 `cat testProcess.pid`
+file="/home/ec2-user/testProcess.pid"
+if [ -f "$file" ]
+then
+	kill -9 `cat testProcess.pid`
+else
+	echo "$file not found."
+fi
